requirements.txt
Django==4.2.7
djangorestframework==3.14.0
spacy==3.7.2
scikit-learn==1.3.2
pdfminer.six==20221105
python-docx==1.1.0
joblib==1.3.2